typecast input to 64 bit int for bit shift macros that shift 32 bits
LVAL(...)<<32 shifts a 32-bit value by 32 — undefined behavior, triggers -Wshift-count-overflow

+++ include/atalk/byteorder.h
@@ -121,7 +121,7 @@ it also defines lots of intermediate macros, just ignore those :-)
 #define SVALS(buf,pos) ((int16)SVAL(buf,pos))
 #define IVAL(buf,pos) (SVAL(buf,pos)|SVAL(buf,(pos)+2)<<16)
 #define IVALS(buf,pos) ((int32_t)IVAL(buf,pos))
-#define LVAL(buf,pos) (IVAL(buf,pos)|IVAL(buf,(pos)+4)<<32)
+#define LVAL(buf,pos) (IVAL(buf,pos)|((uint64_t)IVAL(buf,(pos)+4))<<32)
 #define LVALS(buf,pos) ((int64_t)LVAL(buf,pos))
 
 #define SSVALX(buf,pos,val) (CVAL_NC(buf,pos+1)=(unsigned char)((val)&0xFF),CVAL_NC(buf,pos)=(unsigned char)((val)>>8))
@@ -193,7 +193,7 @@ it also defines lots of intermediate macros, just ignore those :-)
 /* now the reverse routines - these are used in nmb packets (mostly) */
 #define SREV(x) ((((x)&0xFF)<<8) | (((x)>>8)&0xFF))
 #define IREV(x) ((SREV(x)<<16) | (SREV((x)>>16)))
-#define LREV(x) ((IREV(x)<<32) | (IREV((x)>>32)))
+#define LREV(x) (((uint64_t)IREV(x)<<32) | (IREV((x)>>32)))
 
 #define RSVAL(buf,pos) SREV(SVAL(buf,pos))
 #define RSVALS(buf,pos) SREV(SVALS(buf,pos))
